+++ app/src/main/java/com/akinci/socialapitrial/feature/secure/userlist/repository/UserListRepository.kt
@@ -7,7 +7,7 @@ import com.akinci.socialapitrial.feature.secure.userlist.data.output.singout.Sig
 
 interface UserListRepository {
     suspend fun signOut(): Resource<SignOutResponse>
-    suspend fun fetchFollowers(cursor : Int): Resource<FollowerOrFriendResponse>
-    suspend fun fetchFollowings(cursor : Int): Resource<FollowerOrFriendResponse>
+    suspend fun fetchFollowers(cursor : Long): Resource<FollowerOrFriendResponse>
+    suspend fun fetchFollowings(cursor : Long): Resource<FollowerOrFriendResponse>
     suspend fun getUserInfo(userId: Int, userName: String): Resource<UserResponse>
 }
